+++ verik-compiler/src/main/kotlin/io/verik/compiler/serialize/source/ExpressionSerializer.kt
@@ -185,6 +185,10 @@ object ExpressionSerializer {
         serializeContext.append("super")
     }
 
+    fun serializeBreakStatement(serializeContext: SerializeContext) {
+        serializeContext.appendLine("break;")
+    }
+
     fun serializeReturnStatement(returnStatement: EReturnStatement, serializeContext: SerializeContext) {
         val expression = returnStatement.expression
         if (expression == null) {