+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Support string concatenation.
 - Support associative array `set` and `get`.
 - Support `MailBox` class.
+- Support break statements.
 
 ## [0.1.14]
 ### Importer